Move store-settings to config
