+++ config/store/categories.js
